Ensure all preview uses of React 18 API are shimmed
