+++ code/core/src/shared/open-service/sync-test/local-command/local-command.stories.tsx
@@ -11,6 +11,7 @@ import { localCommandSyncService } from './preview.ts';
 const store = createDemoStore('');
 
 function LocalCommandDemo() {
+  // Safe to use React 18 API because this is only loaded in our own UI, not in React sandboxes.
   const value = useSyncExternalStore(store.subscribe, store.get, store.get);
 
   return (
